+++ CHANGES.md
@@ -23,6 +23,7 @@ New Features:
 - io.SerialPort has a new method send_line
 - geometry module has convenience functions for predefined geometrical
   shapes used for plotting with stimuli.Shape 
+- geometry.lines_intersection_point calculates interscetion point of two lines
 - stimuli.Line is internally based on shape and has the new methods
   get_shape and get_connected_shape